\name{DSM_HieroglyphsMatrix}
\alias{DSM_HieroglyphsMatrix}
\docType{data}
\title{
  A Small Co-occurrence Matrix (wordspace)
}
\description{

  A small co-occurrence matrix of verb-object combinations from the British National Corpus (BNC)
  Verbs correspond to columns of the matrix and their object nouns to rows.
  This matrix is shown as the "hieroglyphs" example in the DSM turorial.

}
\usage{

DSM_HieroglyphsMatrix

}
\format{
  
  A numeric matrix with 7 rows and 6 columns.
  
}
\details{

  TODO
  
}
\source{

  TODO
  
}
\references{

  TODO
  
}
\examples{

## TODO: illustrate basic DSM operations with this matrix

}
\keyword{datasets}
