\name{DSM_TermTermMatrix}
\alias{DSM_TermTermMatrix}
\docType{data}
\title{
  Example of a Term-Term Co-occurrence Matrix (wordspace)
}
\description{

  This matrix is a typical example of a term-term DSM co-occurrence matrix,
  obtained from the English Wikipedia.

}
\usage{

DSM_TermTermMatrix

}
\format{

  A numeric matrix with 7 rows and 7 columns.

}
\details{

  TODO
  
}
\source{

  TODO: WaCkypedia corpus
  
}
\seealso{

  This matrix covers the same target terms as the term-context matrix \code{\link{DSM_TermContextMatrix}}.

}
\examples{
## TODO
}
\keyword{datasets}
