\name{wordspace-package}
\alias{wordspace-package}
\alias{wordspace}
\docType{package}
\title{
  Distributional Semantic Models in R (wordspace)
}
\description{

  This package aims to provide a toy laboratory for research and experimentation in distributional semantics
  as well as a user-friendly environment for building and applying distributional semantic models (DSM) in R 

}
%\details{
%}
\author{

  Stephanie Evert (\url{https://purl.org/stephanie.evert})

}
\references{

  Please cite this package as

  Evert, Stefan (2014). Distributional semantics in R with the wordspace package. In \emph{Proceedings of COLING 2014, the 25th International Conference on Computational Linguistics: System Demonstrations}, pages 110--114, Dublin, Ireland.

  and link to the package homepage at \url{http://wordspace.r-forge.r-project.org/}.

  Tutorial materials using the package for examples and exercises are available from \url{http://wordspace.collocations.de/} under an open-source license.  You will also find some pre-compiled distributional semantic models (DSM) there.

}
\seealso{

  If you are new to the \bold{wordspace} package, you should start by reading the package vignette and trying out the code examples shown there.

  Good starting points into the package documentation are \code{\link{dsm}}, \code{\link{read.dsm.triplet}}, \code{\link{dist.matrix}} and \code{\link{nearest.neighbours}}.

}
%% \examples{
%% }
\keyword{package}
