\name{dim.dsm}
\alias{dim.dsm}
\title{
  Dimensions of a DSM Object (wordspace)
}
\description{

  Retrieve the dimensions of the co-occurrence and/or score matrix 
  represented by a DSM object.

}
\usage{

\method{dim}{dsm}(x)

}
\arguments{

  \item{x}{
    an object of class \code{dsm}
  }

}
\details{

  Note that an assignment form (\code{dim<-}) for modifying dimensions is not provided.

}
\value{
%% TODO

%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
%% ~put references to the literature/web site here ~
}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
## TODO
}
\author{Stefan Evert (\url{http://purl.org/stefan.evert})}
\keyword{TODO}
