\name{DSM_TermContextMatrix}
\alias{DSM_TermContextMatrix}
\docType{data}
\title{
  Example of a Term-Context Co-occurrence Matrix (wordspace)
}
\description{

  This matrix is a typical example of a term-context DSM co-occurrence matrix,
  obtained from the English Wikipedia.

}
\usage{

DSM_TermContextMatrix

}
\format{

  A numeric matrix with 7 rows and 7 columns.
  
}
\details{

  TODO
  
}
\source{

  TODO: WaCkypedia corpus
  
}
\seealso{

  This matrix covers the same target terms as the term-term matrix \code{\link{DSM_TermTermMatrix}}.

}
\examples{
## TODO
}
\keyword{datasets}
