\name{wordspace-package}
\alias{wordspace-package}
\alias{wordspace}
\docType{package}
\title{
  Distributional Semantic Models in R (wordspace)
}
\description{

  A toy laboratory for research and experimentation in distributional semantics.

  \strong{Important note:} This package is currently in alpha stage.  All code and interfaces should be considered experimental and may change at any time without notice.

}
\details{

  \tabular{ll}{
    Package: \tab wordspace\cr
    Type: \tab Package\cr
    Version: \tab 0.0-3\cr
    Date: \tab 2010-10-27\cr
    License: \tab GPL-3\cr
    LazyLoad: \tab yes\cr
  }

}
\author{
  Stefan Evert (\url{http://purl.org/stefan.evert})
}
\references{
}
\seealso{
  \url{http://wordspace.collocations.de}
}
\examples{
}
\keyword{package}
