\name{wordspace-package}
\alias{wordspace-package}
\alias{wordspace}
\docType{package}
\title{
What the package does (short line)
}
\description{
More about what it does (maybe more than one line)
}
\details{
\tabular{ll}{
Package: \tab wordspace\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-09-12\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}
}
\author{
Who wrote it

Maintainer: Who to complain to <yourfault@somewhere.net>
}
\references{
}
\keyword{ package }
\seealso{
E.g. \code{\link[<pkg>:<pkg>-package]{<pkg>}}
}
\examples{
}
